Turso Integrated
Turso was integrated via the logger module. The idea is to log every
outgoing message into a turso database. The inteface is simple
inititalize the turso connection, initialize the LoggerContext and use
the log_to_turso() function to send the LoggerContext to Turso.

I refactored the main function and moved some initialization functions
to seperate functions: init_sentry() and init_turso().

Added the database schema that pertains to the logging database.

Ran cargo update to updated dependencies.