Enabled CI/CD Cache
The target folder created via the rust compilation process is cached.
I'm trying to work arround some of the gitlab container limitations when
it comes to cargo. The cargo registry can't be cached using the
traditinal cache method; you can only cache content withing the same
CWD. So I resorted to archiving the the cargo registry with rsync inside
the CWD. This registry folder is then cached using the regular cache
method to carry it over to then next job. Once on the new job, I use
rsync again to move that content the current cargo registry in the
running image.